// src/types/Calculator/Template.ts
export interface Template {
  amount: number
  enabled: number
  material: number
  price: number
  sale_price: number
  subtype: string
  type: string
  warranty: number

  AnglesCount?: number
  ArchType?: number
  BackWrap?: number
  BandColor?: number
  BlockCount?: number
  BracketComplect?: number
  Color?: number
  ConnectorCorner?: number
  ConnectorUni?: number
  ControlDevice?: number
  ControlHeight?: number
  ControlType?: number
  ControlSide?: number
  CorrectorsCount?: number
  DownPlumb?: number
  Drive?: number
  EndBracket?: number
  Fillet?: number
  HasBox?: number
  InsertCount?: number
  InsertSize?: number
  InstallPlace?: number
  Lambreken?: number
  LamelsMaterial?: number
  Magnet?: number
  MaterialCalc?: number
  MountPoint?: number
  MountType?: number
  MountTypeStatic?: number
  Naborka?: number
  promoCode?: string
  RingsAmount?: number
  cost_height?: number
  cost_width?: number
  gab_height?: number
  gab_width?: number
  height?: number
  karniz?: object
  old_price?: number
  perc?: number
  width?: number
}